Added my gravity and orb code
I basically copied over what I did on my own project. I deleted the main ViewController, as it was obsolete. I rearranged the files. I deleted obsolete code in GameScene.swift and GameViewController.swift.

I used the Sprite Kit Scene to add in the ball and the barriers, and then programmed the gravity and acceleration controls in GameScene.swift.